+++ crypto/keyring/keyring.go
@@ -518,11 +518,11 @@ func (ks keystore) Delete(uid string) error {
 func (ks keystore) KeyByAddress(address []byte) (*Record, error) {
 	ik, err := ks.db.Get(addrHexKeyAsString(address))
 	if err != nil {
-		return nil, wrapKeyNotFound(err, "key with given address not found") // we do not print the address for not needing an address codec
+		return nil, wrapKeyNotFound(err, fmt.Sprint("key with address ", address, " not found"))
 	}
 
 	if len(ik.Data) == 0 {
-		return nil, wrapKeyNotFound(err, "key with given address not found") // we do not print the address for not needing an address codec
+		return nil, wrapKeyNotFound(err, fmt.Sprint("key with address ", address, " not found"))
 	}
 
 	return ks.Key(string(ik.Data))